our select works!

+++ models/hotel.js
@@ -5,8 +5,9 @@ var Hotel = db.define('hotel', {
     name: {
         type: Sequelize.STRING
     },
-    num_stars:{
-        type: Sequelize.INTEGER,
+
+    num_stars: {
+        type: Sequelize.DECIMAL,
         validate: {
             min: 1,
             max: 5
@@ -14,13 +15,13 @@ var Hotel = db.define('hotel', {
     },
     amenities: {
         type: Sequelize.STRING
-        // set: function(amenities) {
-        //     amenities = amenities || [];
-        //     if (typeof amenities === 'string') {
-        //         amenities = amenities.split(",")
-        //     }
-        // this.setDataValue('amenities', amenities)
-        // }
+            // set: function(amenities) {
+            //     amenities = amenities || [];
+            //     if (typeof amenities === 'string') {
+            //         amenities = amenities.split(",")
+            //     }
+            // this.setDataValue('amenities', amenities)
+            // }
     }
 })
